Update Web and API

+++ admin/candidates_delete.php
@@ -3,13 +3,26 @@
 
 	if(isset($_POST['delete'])){
 		$id = $_POST['id'];
-		$sql = "DELETE FROM candidates WHERE id = '$id'";
-		if($conn->query($sql)){
-			$_SESSION['success'] = 'Candidate deleted successfully';
-		}
-		else{
-			$_SESSION['error'] = $conn->error;
-		}
+		
+		$sqlgambar = "SELECT photo FROM candidates WHERE id = '$id'";
+			$query = $conn->query($sqlgambar);
+			$row = $query->fetch_assoc();
+			$nama_file = "../images/$row[photo]";
+			
+			if (file_exists($nama_file)) {
+				 if (unlink($nama_file)) {
+					$sql = "DELETE FROM candidates WHERE id = '$id'";
+					if($conn->query($sql)){
+						$_SESSION['success'] = 'Candidate deleted successfully';
+						}else{
+						$_SESSION['error'] = $conn->error;
+					}
+				} else {
+					$_SESSION['error'] = $conn->error;
+				}
+				
+			}
+		
 	}
 	else{
 		$_SESSION['error'] = 'Select item to delete first';
